requirements.txt
molo.core==6.8.3
molo.surveys==6.7.5
molo.commenting==6.2.9
molo.pwa==6.2.0
molo.servicedirectory==6.1.0
molo.yourwords==6.3.0
molo.polls==6.3.0
mote-prk==0.2.1
dateutils==0.6.6
celery<4.0
Unidecode==0.04.16
gunicorn
psycopg2
django_compressor==2.2
django-mptt==0.8.7
django-google-analytics-app==4.2.0
django-storages
boto
django-modelcluster==3.1
mozilla_django_oidc==1.0.0
# Required by ImageHash, install a working version
scipy==0.19.1
# need to remove ES from here when we add it to Molo.core
elasticsearch>=2.0.0,<3.0.0
